tests/wpt/web-platform-tests/css/css-typed-om/the-stylepropertymap/inline/get.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>StylePropertyMap.get tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#get-a-value-from-a-stylepropertymap">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<body>
<script>
'use strict';

test(t => {
  const styleMap = createInlineStyleMap(t);
  assert_throws(new TypeError(), () => styleMap.get('lemon'));
}, 'Calling StylePropertyMap.get with an unsupported property throws a TypeError');

test(t => {
  const styleMap = createInlineStyleMap(t);
  assert_equals(styleMap.get('height'), null);
}, 'Calling StylePropertyMap.get with a property not in the property model returns null');

test(t => {
  const styleMap = createInlineStyleMap(t, '--foo: auto');
  assert_equals(styleMap.get('--Foo'), null);
}, 'Calling StylePropertyMap.get with a custom property not in the property model returns null');

test(t => {
  const styleMap = createInlineStyleMap(t, 'width: 10px; height: 20px');
  assert_style_value_equals(styleMap.get('width'), CSS.px(10));
}, 'Calling StylePropertyMap.get with a valid property returns the correct entry');

test(t => {
  const styleMap = createInlineStyleMap(t, 'height: 20px; width: 10px;');
  assert_style_value_equals(styleMap.get('wIdTh'), CSS.px(10));
}, 'StylePropertyMap.get with a valid property in mixed case returns the correct entry');

test(t => {
  const styleMap = createInlineStyleMap(t, '--foo: auto; --bar: 10px');
  assert_style_value_equals(styleMap.get('--foo'), new CSSUnparsedValue(' auto'));
}, 'Calling StylePropertyMap.get with a valid custom property returns the correct entry');

test(t => {
  const styleMap = createInlineStyleMap(t, 'width: 10px; transition-duration: 1s, 2s; height: 10px;');
  assert_style_value_equals(styleMap.get('transition-duration'), CSS.s(1));
}, 'Calling StylePropertyMap.get with a list-valued property returns only the first value');

</script>
